fedora/scripts/requirements.txt
attrs
certifi
charset-normalizer
colorama
coloredlogs
flatbuffers
humanfriendly
idna
imageio
jsonschema
jsonschema-specifications
lazy_loader
llvmlite
mpmath
networkx
numba
numpy
onnxruntime
opencv-python-headless
packaging
pillow
platformdirs
pooch
protobuf
PyMatting
PyMuPDF
PyPDF2
pyreadline3
referencing
rembg
requests
rpds-py
scikit-image
scipy
sympy
tifffile
tqdm
urllib3
